Split track tile into components
